feat(testing): add force_python_pump_fallback utility for stream tests
Introduce force_python_pump_fallback in cuprum._testing to patch stream dispatch internals during tests. This utility forces fallback to the Python pump implementation and counts calls, enabling more granular testing of stream backend behavior when FD extraction fails. Update related tests to use this new helper.
